Update README.md
I wanted to add two libraries to the list but didnt see a section that suited them well.
Maybe adding a "Text" section would be a good idea if it is ok to add these two libraries to the list since they are not strictly "UI".

How does it look? Am I far off the mark?

Thanks!